taskapp/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-filter==2.4.0
django-simple-history==2.11.0
djangorestframework==3.11.2
djangorestframework-jsonapi==3.2.0
inflection==0.5.1
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1